chore(whatsapp): remove legacy transport vestiges
