+++ apps/web/src/hosted/v2/channels/channel-detail-page.logic.test.ts
@@ -94,7 +94,6 @@ describe("nativeTransportSummary", () => {
 				mode: "none",
 				reason: "shared-bot-transport-unavailable",
 				supportsOutboundMessages: false,
-				supportsRawRelay: false,
 			}),
 		).toEqual({
 			status: "Unavailable",
