+++ backend/app/services/whatsapp_sidecar_registry.py
@@ -2,15 +2,13 @@ from __future__ import annotations
 
 import asyncio
 import logging
-from collections.abc import Callable, Mapping
-from dataclasses import replace
+from collections.abc import Callable
 from typing import Protocol
 from uuid import UUID
 
 from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.core.config import settings
 from app.core.database import async_session_factory
 from app.models.channel import (
     CHANNEL_PROVIDER_WHATSAPP,
@@ -20,6 +18,10 @@ from app.models.channel import (
     WHATSAPP_ONBOARDING_STATE_ERROR,
     ChannelAccount,
     ChannelWhatsAppOnboardingSession,
+)
+from app.services.whatsapp_delivery_transport import (
+    DEFAULT_WHATSAPP_SIDECAR_SOCKET_PATH,
+    configured_whatsapp_sidecar_session_id,
 )
 from app.services.whatsapp_native_transport import (
     WhatsAppBaileysSidecarClient,
@@ -42,8 +44,6 @@ from app.services.whatsapp_provider_bridge import (
 )
 
 log = logging.getLogger(__name__)
-
-DEFAULT_WHATSAPP_SIDECAR_SOCKET_PATH = "/run/clawdi-whatsapp/sidecar.sock"
 _UNRELEASED_STATES = (
     "generating",
     "ready",
@@ -84,48 +84,6 @@ class WhatsAppSidecarClient(WhatsAppNativeUpstreamClient, Protocol):
 
 SidecarClientFactory = Callable[[WhatsAppBaileysSidecarConfig], WhatsAppSidecarClient]
 _active_registry: ConfiguredWhatsAppSidecarRegistry | None = None
-_delivery_sidecar_service: WhatsAppBaileysSidecarService | None = None
-
-
-def resolve_whatsapp_delivery_transport(
-    account: ChannelAccount,
-) -> WhatsAppProviderTransportAdapter | None:
-    if account.provider != CHANNEL_PROVIDER_WHATSAPP:
-        return None
-    config = account.config if isinstance(account.config, dict) else {}
-    connection_mode = config.get("connection_mode")
-    if connection_mode == "baileys_managed":
-        session_id = account.id
-    elif connection_mode == "baileys_custom":
-        session_id = _configured_session_id(config)
-    else:
-        return None
-    if session_id is None:
-        return None
-
-    service_config = _configured_delivery_service()
-    if service_config is None:
-        return None
-    session_config = replace(service_config, account_id=session_id)
-    if config.get("sidecar_config_revision") != session_config.binding_revision:
-        return None
-
-    global _delivery_sidecar_service
-    if _delivery_sidecar_service is None:
-        _delivery_sidecar_service = WhatsAppBaileysSidecarService(service_config)
-    return WhatsAppProviderTransportAdapter(_delivery_sidecar_service.session_client(session_id))
-
-
-def _configured_delivery_service() -> WhatsAppBaileysSidecarConfig | None:
-    api_token = settings.channel_whatsapp_baileys_sidecar_token.get_secret_value().strip()
-    if not api_token:
-        return None
-    base_url = settings.channel_whatsapp_baileys_sidecar_url.strip() or None
-    return WhatsAppBaileysSidecarConfig(
-        api_token=api_token,
-        base_url=base_url,
-        unix_socket_path=None if base_url else DEFAULT_WHATSAPP_SIDECAR_SOCKET_PATH,
-    )
 
 
 class ConfiguredWhatsAppSidecarRegistry:
@@ -481,7 +439,7 @@ class ConfiguredWhatsAppSidecarRegistry:
                 config = account.config if isinstance(account.config, dict) else {}
                 if config.get("connection_mode") != "baileys_custom":
                     continue
-                session_id = _configured_session_id(config)
+                session_id = configured_whatsapp_sidecar_session_id(config)
                 revision = config.get("sidecar_config_revision")
                 if session_id is None or not isinstance(revision, str) or session_id in owners:
                     if session_id is not None:
@@ -678,11 +636,3 @@ class ConfiguredWhatsAppSidecarRegistry:
 
 def get_active_whatsapp_sidecar_registry() -> ConfiguredWhatsAppSidecarRegistry | None:
     return _active_registry
-
-
-def _configured_session_id(config: Mapping[str, object]) -> UUID | None:
-    raw = config.get("sidecar_account_id")
-    try:
-        return UUID(raw) if isinstance(raw, str) else None
-    except ValueError:
-        return None
